requirements.txt
discord.py
discord.py[voice]
easy_pil
requests
python-dotenv